Fix for “Error: TypeError: path must be a string”
Fix for “Error: TypeError: path must be a string” when using requirejs
(removedCombined = true) during compile

+++ hbs.js
@@ -576,7 +576,7 @@ define([
             });
           }
 
-          if ( config.removeCombined ) {
+          if ( config.removeCombined && path ) {
             fs.unlinkSync(path);
           }
